Reworked performance multiplayer for song loading.
A ready status property was added which is being used to indicate that the players aren't ready for the match to start till their system fully loads the song. After that the ready status of the user is updated and synchronized with the rest of the clients. When all clients are ready a play/seek to the start of the song is performed. This allows us to work on the song loader to load the songs remotely and wait for all players to load the song.
The ready status is an enum that changes from 0 to 100. So if values are defind in between it can be used as a progress bar value as well.
Negative values can be added to indicate non progress based non-readiness.
UI Indicators for the ready status should be added in the future so that the users will be aware why the game has not started yet.