New Audio Framing mechanism
A new audio framing mechanism has been implemented both in the module and the game.

Now audio frames will be triggered by a deferred signal emission from the audio effect instance. This will help avoid analyzing duplicate frames or missing frames. Due to this the Note recognition is more deterministic and accurate.

Version for the gitub action was bumped to use the newest custom engine build.